backend/requirements.txt
flask
flask-cors
pandas
tabulate
python-dotenv
mysql-connector-python
sqlparse
google-generativeai
requests
sqlalchemy
pymysql
